[REFACTOR] - Change in the form of the `methods` module
[*] Transition from a ‘folder’ module to a ‘single file’ module for the `methods` module
[-] Removal of `methods/*.rs` files

[ ] All tests are functional
[ ] Related to issue #8
